Remove dependency on ding_list module
With the update to Ding 6.0.0 the ding_list module is now obsolete 
and should no longer be a dependency.

Buttons with multiple actions are no longer supported in Ding 6 but 
we still need the functionality here to display multiple libraries in
a more intelligent manner. Consequently we reuse the classes and 
include the JS and compiled CSS from Ding 5.0.1.

Ref. KBIB-333