Make contact bottom block translatable
I am doing the minimum here. The footer block uses some
of the same variables as the contact page. Those variables
are exposed in the admin page. The contact page (and the admin page)
er not translatable - So in order to make the footer work,
I expose the variables needed for that in the contact admin page.

Ie "DA: Section title" and "EN: Section title". Furthermore not all
variables on the admin page are translatable, which could cause some confusion.

KKB-340