Handle and expose exposed dates
If the to dates have been saved earlier on, and are removed in the current save loop,
we need to programmatically unset them before saving the note. Otherwise they live
on through the save loop, and still be visible in the kultunaut_feed (they will
not be visible in the node form).

KKB-346