Move wayfinder assets from #attach to drupal_add_*. KKB-466
For some ungodly reason, Drupal refuses to add
assets when they're being attached through AJAX.
Before this commit, the wayfinder worked on URLs like
ting/object/870970-basis%3A46687752 where the assets were
added on page-load, but it was broken on pages where
entities were loaded with AJAX such as lists.

I've tried a million things and nothing works, so this
is the ultimate hammer: Addings the assets on all pages.
I've checked the assets, and they're not gigantic, so
it's not the end of the world.