Set wayfinder content width based on its parent slick-list width
We have to use JS because we cannot set the width of the content to 100% with css.
This will only make it the width of its parent, which is a slick slide and is about 260px.

We could set it to its maximum of 1124px, but if the use resizes or has a screen that is smaller
than that, it wont work.

KKB-491