Make kkb_help accordions accessible
The accordions are enabled by misc/collapse.js, that adds some
elements to the dom - fx a.fieldset-title. Since this is the
element that we want to manipulate, we need to observe the dom
and wait for the elements to be inserted.

Once they are there we can set the title as the aria-expanded controller
and the wrappers as the controlled region.

Unfortunately we need to make the fieldset-title a tag into a button in order
for screen readers to read it out correctly. This change leads to another
fix, where we have to call toggleFieldset in collapse.js manually.

KKB-489